Tests/LibWeb/Text/expected/html/activeElement-body.txt

document.activeElement === document.body: true
document.activeElement is null: false
document.activeElement tagName: BODY
